fix: restructuring for statuses alone feat: adding pipenv structures
