+++ Pipfile
@@ -1,0 +1,19 @@
+[[source]]
+url = "https://pypi.org/simple"
+verify_ssl = true
+name = "pypi"
+
+[packages]
+dataengineeringutils3 = "*"
+arrow-pd-parser = "*"
+numpy = "*"
+pandas = "*"
+requests = "*"
+pydbtools = "*"
+data-linter = "*"
+
+[dev-packages]
+ipykernel = "*"
+
+[requires]
+python_version = "3.9"
