+++ python_scripts/functions/api_requests.py
@@ -1,118 +1,23 @@
 import pandas as pd
 import requests
-import re
-import json
 from arrow_pd_parser import writer
-import python_scripts.s3_utils as s3_utils
 from mojap_metadata import Metadata
 
-from python_scripts.constants import (
-    land_location,
+from python_scripts.functions.api_helpers import (
+    get_payload,
+    make_booking_params,
+    matrix_authenticate,
+    extract_locations,
+    fix_faulty_time_col,
+    camel_to_snake_case,
 )
-from python_scripts.column_renames import (
-    location_renames,
-    bookings_renames,
-)
+from python_scripts.constants import land_location
+
+from python_scripts.column_renames import bookings_renames, location_renames
 
 from logging import getLogger
 
 logger = getLogger(__name__)
-
-
-def read_json(file_path: str) -> dict:
-    """Reads a json file in as a dictionary
-
-    Parameters
-    ----------
-    file_path :
-        file path of the JSON to read from
-
-    Returns
-    -------
-        dictionary representing the json file
-    """
-    f = open(file_path)
-    return json.loads(f.read())
-
-
-def get_secrets() -> dict:
-    return s3_utils.read_json_from_s3("alpha-dag-matrix/api_secrets/secrets.json")
-
-
-def matrix_authenticate(session: requests.Session) -> requests.Session:
-    # Add this before the error occurs
-    for handler in logger.handlers:
-        for record in handler.records:
-            print(record.__dict__)
-    secrets = get_secrets()
-    username = secrets["username"]
-    password = secrets["password"]
-
-    url = "https://app.matrixbooking.com/api/v1/user/login"
-    session.post(url, json={"username": username, "password": password})
-    return session
-
-
-def get_booking_categories(session: requests.Session) -> pd.DataFrame:
-    """Returns pandas dataframe containing all booking categories
-    that are available to organisation
-
-     Parameters:
-    session (requests.sessions.Session): Authenticated session to
-        matrix booking API
-    """
-
-    # Booking categories API url
-    url_booking_cats = "https://app.matrixbooking.com/api/v1/category"
-
-    # Make request and create dataframe
-    res = requests.get(url_booking_cats, cookies=session.cookies).json()
-    df_booking_categories = pd.json_normalize(res)
-
-    return df_booking_categories
-
-
-def make_booking_params(
-    time_from: str,
-    time_to: str,
-    pageSize: int = None,
-    pageNum: int = 0,
-) -> dict:
-    params = {
-        "f": time_from,
-        "t": time_to,
-        "include": "audit",
-        "pageSize": pageSize,
-        "pageNum": pageNum,
-    }
-    return params
-
-
-def get_payload(session, url, parameters):
-    resp = session.get(url=url, cookies=session.cookies, params=parameters)
-    logger.debug(f"GET {resp.url}")
-    logger.debug(f"response status code: {resp.status_code}")
-    return resp.json()
-
-
-def split_s3_path(s3_path: str) -> tuple[str]:
-    """Splits an s3 file path into a bucket and key
-
-    Parameters
-    ----------
-    s3_path :
-        The full (incl s3://) path of a file.
-
-    Returns
-    -------
-        Tuple of the bucket name and key (file path) within that bucket.
-    """
-    if s3_path[:2] != "s3":
-        raise ValueError("S3 file path should start with 's3://'.")
-    path_split = s3_path.split("/")
-    bucket = path_split[2]
-    key = "/".join(path_split[3:])
-    return bucket, key
 
 
 def scrape_days_from_api(
@@ -181,15 +86,20 @@ def scrape_days_from_api(
     return raw_bookings
 
 
-def camel_to_snake_case(input_str: str) -> str:
-    # Using regular expressions to find positions with capital letters and insert underscores
-    s1 = re.sub("(.)([A-Z][a-z]+)", r"\1_\2", input_str)
-    s2 = re.sub("([a-z0-9])([A-Z])", r"\1_\2", s1)
-
-    # Handle the case where multiple uppercase letters are present
-    snake_case_str = re.sub("([a-z])([A-Z]+)", r"\1_\2", s2).lower()
-
-    return snake_case_str
+def scrape_locations_from_api(start_date: str) -> pd.DataFrame:
+    ses = requests.session()
+    matrix_authenticate(ses)
+    params = {"f": start_date, "t": "eod"}
+    url = "https://app.matrixbooking.com/api/v1/org/43/locations"
+    logger.info("Scraping locations info")
+    raw_locations = get_payload(ses, url, params=params)
+    unnest_locs = extract_locations(raw_locations)
+    raw_unpacked_locations = (
+        pd.json_normalize(unnest_locs, sep="_")
+        .drop(columns=["locations"])
+        .rename(mapper=camel_to_snake_case, axis="columns")
+    )
+    return raw_unpacked_locations
 
 
 def rename_df(df: pd.DataFrame, renames: dict) -> pd.DataFrame:
@@ -221,31 +131,6 @@ def rename_df(df: pd.DataFrame, renames: dict) -> pd.DataFrame:
     return df
 
 
-def fix_faulty_time_col(df, col):
-    column = df[col].copy()
-    # Check for missing parts (seconds, minutes, hours, microseconds)
-    missing_parts = column.apply(
-        lambda x: (pd.notna(x) and (len(str(x).split(":")) < 3 or "." not in str(x)))
-    )
-
-    def format_timestamp(raw_string):
-        # Generate dynamic format based on missing parts
-        num_parts = len(raw_string.split(":"))
-
-        format_str = (
-            "%Y-%m-%dT"
-            + ":".join(["%H", "%M", "%S"][:num_parts])
-            + (".%f" if "." in raw_string else "")
-        )
-
-        return pd.to_datetime(raw_string, format=format_str).strftime(
-            "%Y-%m-%dT%H:%M:%S.%f"
-        )
-
-    column.loc[missing_parts] = column.loc[missing_parts].apply(format_timestamp)
-    return column
-
-
 def fix_faulty_time_cols(df):
     """_summary_
 
@@ -262,9 +147,7 @@ def fix_faulty_time_cols(df):
     return df
 
 
-def write_raw_data_to_s3(
-    bookings: pd.DataFrame, locations: pd.DataFrame, start_date: str, env: str
-):
+def write_raw_data_to_s3(df: pd.DataFrame, renames: dict, raw_loc: str, env: str):
     """_summary_
 
     Parameters
@@ -276,22 +159,22 @@ def write_raw_data_to_s3(
     start_date : _type_
         _description_
     """
-    raw_bookings_loc = f"{land_location}/bookings/{start_date}/raw-{start_date}.jsonl"
-    raw_locations_loc = f"{land_location}/locations/{start_date}/raw-{start_date}.jsonl"
-    bookings = rename_df(bookings, bookings_renames)
-    bookings = fix_faulty_time_cols(bookings)
-    locations = rename_df(locations, location_renames)
+    df = rename_df(df, renames)
+    df = fix_faulty_time_cols(df)
     writer.write(
-        bookings,
-        raw_bookings_loc,
-    )
-    writer.write(
-        locations,
-        raw_locations_loc,
+        df,
+        raw_loc,
     )
     logger.info(f"{env}: raw booking and location data written to {land_location}.")
 
 
-def scrape_and_write_raw_data(start_date, env):
-    bookings, locations = scrape_days_from_api(start_date, "eod")
-    write_raw_data_to_s3(bookings, locations, start_date, env)
+def scrape_and_write_raw_bookings_data(start_date, env):
+    raw_bookings_loc = f"{land_location}/bookings/{start_date}/raw-{start_date}.jsonl"
+    bookings = scrape_days_from_api(start_date, "eod")
+    write_raw_data_to_s3(bookings, bookings_renames, raw_bookings_loc, env)
+
+
+def scrape_and_write_raw_locations_date(start_date, env):
+    raw_locations_loc = f"{land_location}/locations/{start_date}/raw-{start_date}.jsonl"
+    locations = scrape_locations_from_api(start_date)
+    write_raw_data_to_s3(locations, location_renames, raw_locations_loc, env)
